// src/pocAngularOfflineServer/src/model/hockeyPlayer.ts
/**
 * pocAngularOfflineServer
 * This is the API specification of the server
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface HockeyPlayer {
    id?: string;
    firstname?: string;
    lastname?: string;
}